image_colorization/requirements.txt
tensorflow~=2.4.1
Keras~=2.4.3
numpy~=1.19.5
matplotlib~=3.3.4
scikit-image~=0.18.1
Pillow~=8.1.0
scikit-learn~=0.24.2
tqdm~=4.60.0